Deleted unuseful div, files named without camelCase
