+++ src/view/new-without-destination.js
@@ -1,7 +1,7 @@
 import { currentTime } from '../utils/utils.js';
-import { TYPES, CITIES } from '../mock/createData.js';
-import { addOffers, createTypes, createCities, getFormatTime } from '../utils/renderingUtils.js';
-import { createElement } from '../view/renderingUtils.js';
+import { TYPES, CITIES } from '../mock/create-data.js';
+import { addOffers, createTypes, createCities, getFormatTime } from '../utils/rendering-utils.js';
+import { createElement } from './rendering-utils.js';
 
 const addNewPointWithoutDestination = (points = {}) => {
   const {
