+++ src/view/rendering-utils.js
