+++ src/view/trip-list-ul.js
@@ -1,13 +1,13 @@
-import { createElement } from '../view/renderingUtils.js';
+import { createElement } from './rendering-utils.js';
 
-const tripListLi = () => '<li class="trip-events__item"></li>';
-export default class TripListLi {
+const tripListUl = () => '<ul class="trip-events__list"></ul>';
+export default class TripListUl {
   constructor() {
     this._element = null;
   }
 
   getTemplate() {
-    return tripListLi();
+    return tripListUl();
   }
 
   getElement() {
